requirements.txt
Flask==2.2.3
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
scikit-image==0.20.0
torchvision==0.14.1
waitress==2.1.2


